requirements.txt
pywebview
pyobjc
requests